[CDG-66][CDG-67] - Refactor relation list component
